src/AppLogistics.Controllers/Configuration/Sectors/SectorsController.cs
using AppLogistics.Components.Mvc;
using AppLogistics.Objects;
using AppLogistics.Services;
using AppLogistics.Validators;
using Microsoft.AspNetCore.Mvc;

namespace AppLogistics.Controllers.Configuration
{
    [Area("Configuration")]
    public class SectorsController : ValidatedController<ISectorValidator, ISectorService>
    {
        public SectorsController(ISectorValidator validator, ISectorService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create([BindExcludeId] SectorView sector)
        {
            if (!Validator.CanCreate(sector))
            {
                return View(sector);
            }

            Service.Create(sector);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(int id)
        {
            return NotEmptyView(Service.Get<SectorView>(id));
        }

        [HttpGet]
        public ActionResult Edit(int id)
        {
            return NotEmptyView(Service.Get<SectorView>(id));
        }

        [HttpPost]
        public ActionResult Edit(SectorView sector)
        {
            if (!Validator.CanEdit(sector))
            {
                return View(sector);
            }

            Service.Edit(sector);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(int id)
        {
            return NotEmptyView(Service.Get<SectorView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(int id)
        {
            if (!Validator.CanDelete(id))
            {
                return RedirectToAction("Delete", new { id });
            }

            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
